Move scenario FX to strata-data
Move FX provider and rename `CurrencyConvertible`
