+++ modules/calc/src/main/java/com/opengamma/strata/calc/result/DefaultScenarioResult.java
@@ -32,10 +32,10 @@ import org.joda.beans.impl.direct.DirectMetaPropertyMap;
 import com.google.common.collect.ImmutableList;
 import com.opengamma.strata.basics.currency.Currency;
 import com.opengamma.strata.basics.currency.FxConvertible;
-import com.opengamma.strata.calc.runner.CurrencyConvertible;
-import com.opengamma.strata.calc.runner.ScenarioFxRateProvider;
 import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.collect.Messages;
+import com.opengamma.strata.data.scenario.ScenarioFxConvertible;
+import com.opengamma.strata.data.scenario.ScenarioFxRateProvider;
 
 /**
  * A scenario result holding one value for each scenario.
@@ -47,7 +47,7 @@ import com.opengamma.strata.collect.Messages;
  */
 @BeanDefinition(builderScope = "private")
 public final class DefaultScenarioResult<T>
-    implements ScenarioResult<T>, CurrencyConvertible<ScenarioResult<?>>, ImmutableBean, Serializable {
+    implements ScenarioResult<T>, ScenarioFxConvertible<ScenarioResult<?>>, ImmutableBean, Serializable {
 
   /**
    * The calculated values, one per scenario.
