+++ modules/calc/src/test/java/com/opengamma/strata/calc/result/SingleScenarioResultTest.java
@@ -19,9 +19,9 @@ import org.testng.annotations.Test;
 
 import com.google.common.collect.ImmutableList;
 import com.opengamma.strata.basics.currency.CurrencyAmount;
-import com.opengamma.strata.calc.runner.ScenarioFxRateProvider;
 import com.opengamma.strata.collect.array.DoubleArray;
 import com.opengamma.strata.data.scenario.FxRatesArray;
+import com.opengamma.strata.data.scenario.ScenarioFxRateProvider;
 
 /**
  * Test {@link SingleScenarioResult}.
