+++ modules/data/src/test/java/com/opengamma/strata/data/scenario/CombinedScenarioMarketDataTest.java
@@ -19,9 +19,6 @@ import com.opengamma.strata.collect.timeseries.LocalDateDoubleTimeSeries;
 import com.opengamma.strata.data.FieldName;
 import com.opengamma.strata.data.ObservableId;
 import com.opengamma.strata.data.ObservableSource;
-import com.opengamma.strata.data.scenario.ImmutableScenarioMarketData;
-import com.opengamma.strata.data.scenario.MarketDataBox;
-import com.opengamma.strata.data.scenario.ScenarioMarketData;
 
 @Test
 public class CombinedScenarioMarketDataTest {
